refactor : 패키지 구조 개선
