+++ src/main/java/bowling/domain/framestatus/Empty.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.framestatus;
 
 public class Empty implements FrameStatus {
     @Override
