+++ src/main/java/bowling/domain/framestatus/FrameManagement.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.framestatus;
 
 public class FrameManagement {
