+++ src/main/java/bowling/domain/framestatus/Strike.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.framestatus;
 
 public class Strike implements FrameStatus {
