+++ src/test/java/bowling/domain/PinTest.java
@@ -1,5 +1,6 @@
 package bowling.domain;
 
+import bowling.domain.frame.Pin;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
